requirements.txt

google-genai==1.60.0
python-dotenv==1.2.1

